Vanilla sm and notifications
Added a vanillaSM class to detect the vanilla super metroid ROM, functions as expected, with all the expected bugs

updated the typing of a dummy variable in GameState to avoid a sign/unsign warning

Updated the update_sm_counts() method to notify the player when they recieve missiles, super missiles, power bombs, energy tanks, or reserve tanks. and how many of each.